requirements.txt
tweepy==4.14.0
openai==0.27.8
python-dotenv==1.0.0
